Serialize the event as a hash once for each handler to process.
Previously, we pulled out specific data from the event and threw it on sidekiq
even though the data was specific to the single handler we have right now.

This commit generically puts the entire raw event on sidekiq as a hash so
future workers don't need to do API calls for data that was already in the event.

Additionally, we could have had a N+1 situation if the data we previously
put on the queue did not contain what we needed as each handler may need to
do N+1 API queries to get data that the listener already could have pre-fetched
once.