Create a development Procfile that won't be tracked.
This gets my git status back to clean by ignoring a custom Procfile
that's just a copy of the one checked in. It can now be edited to fit
development needs without creating a dirty git state. The README has
been updated to reflect the change, and includes instructions on running
foreman with a custom Procfile.