Upgrade to RSpec 3.
This seemed like low-hanging fruit and does the bare minimum to upgrade
while keeping all tests passing and removing any deprecation warnings.

The should syntax has been enabled here as it's used indirectly in some
example code for the rubocop checker.

`example` is not available anymore inside of examples, but it can be
accessed through the object yielded in `it`, `before`,
etc.. blocks. Setting an ivar in a before block and using this was a
quick hack to get back to green.