Rehash legacy specs for IssueManager
This commit reworks some of the forgotten/failing tests that resided in
`lib/bot/spec`. An attempt to tweak them in order to get them to green
was made, but owing to the extensive stubbing and setup in general it
became too hard. I decided to rewrite them using the same
specs (i.e. they test the same thing, but differently).

In particularly, it:

* Introduces an `IssueManager.build` method which replaces the `.new`
  method. This allows for the constructor to be much simpler and for a
  lot of the details to be abstracted away, which simplifies the tests
  somewhat.
* Updates everything to use `IssueManager.build` where appropriate.
* Removes VCR. It would be nice to bring this back at some point.
* Avoids stubbing Octokit (object we don't own), and leverages some
  verifying doubles to give more confidence

Some unfortunate things that further refactoring will hopefully eliminate:

* Stubbing `File.write` to avoid changing the world - `IssueManager`
  persists data to a local yaml file.
* Stubbing `YAML.load_file` for the same reason.
* Still too much stubbing and especially nested stubs in the setup for
  `IssueManager` specs
* Reduces test coverage (only in theory - in practice these tests
  weren't being run/maintained) around the network boundary. It would be
  nice to add some integration tests here