Check the bug status before making a comment
Sometimes people make PRs with bad rebases that include many commits
which should not have been in the PR.
This has caused the BugzillaPrChecher to comment on every BZ
referenced in those commits with this new PR link.

This change will not completely solve the issue, but will at least
prevent the bot from adding comments to BZs which obviously already
have fixes (statuses not NEW, ASSIGNED, ON_DEV) and should not be
being referenced in PR commits.