Use GitHub search API to construct backport requests view
Using the GitHub search API, every pending backport for a target branch
name can be identified with a single API request. This change adds a new
dashboard view for backport requests per branch.

The request to GitHub when the page is viewed so there is a small chance
of bumping into rate limits. The search API has it's own rate limits
(30/minute), so it should not affect existing bot behavior and should
not be reached under normal circumstances.