Fix issue where invalid assignees were being sent
Previously, we would filter against the organization members, but this
is invalid.  It just *happened* to work in the main repo, because
everyone was an assignee in the main repo, but when moving out to a
split repo, this is no longer valid.

Additionally, organization member caching was only implemented for this
purpose, so instead this is removed in favor of per-repo caching of
assignees.  Also, the org parameter to the GithubNotificationMonitor was
only there for the org member caching, so it too is removed.