Return average in 1 minute increments
Because we're now recording observations on every Github request, we
don't want to show every single datapoint on the chart. Group by and
average the calls in a 1 minute interval.

Data is displayed only for the last 12 hours because anything larger
just isn't that helpful to look at (checked with enormous sample
dataset) We really need to do Grafana if we want to dig in deep yet
still see large trends at the same time.