Fix limit diff from bot db
The bot is currently getting changes in its git stage because the db was
created from a previous era where Rails put limits in its columns
without you specifying. This migration is a no op in the bot's
perspectice but adds these limits for development environments so the
schema is exactly the same as what the bot currently uses.