Squeeze out near nanosecond precision
Ruby doesn't really do a lot with nanosecond precision, and isn't 100%
accurate with the nanosecond features it has. However, this change
should squeeze out the most accurate possible* measurement of
nanoseconds I can come up with.

This change seems to often detect two of the three requests in the
GithubNotificationMonitor that get called in rapid succession.

*Note that according to Ruby documentation, #nsec is actually a more
accurate measurement of nanoseconds, but I don't see how one actually
uses that to record the full timestamp as it's only the fractional bit
of seconds. For more, see http://ruby-doc.org/core-2.4.0/Time.html#method-i-nsec