Use Faraday HTTP caching
This gem adds a Faraday middleware that caches HTTP responses by checking
expiration and validation of the stored responses.

It uses the following headers to make caching decisions:

* Cache-Control
* Age
* Last-Modified
* ETag
* Expires

We might want to consider utilizing Redis for our application cache, but
I've left the default FileStore cache to use for now. Both FileStore and
Redis work across processes so workers should share the same cache
regardless.

I don't know of a very good way to test this other that just shipping it
and watching closely, which is honestly what I recommend we do. Monitor
the logs and watch hit/misses and make sure the bot is behaving and
doing it's chores. Just like having a kid, right?

...right?