Version control Gemfile.lock
*Breaking News*

This is an application, and there are no reasons that I can think of
between our production and development environments that we can't check
in the Gemfile.lock to ensure parity between those environments and
better manage keeping our development environment up to date, as is the
point of Gemfile.lock.

This Gemfile.lock is the exact one used in our current production
environment and should have no observable effect when deployed at this time.