Integrate OctokitWrappers in to GithubService
The new GithubService covers things much more nicely than the current
OctokitWrappers, which was based very closely on the original GitHubApi
Octokit classes.

The wrappers are still quite a useful abstraction. There's too much info
used in responses that is frequently accessed and the wrappers provide a
nice abstracted way of getting to it without caring how. So, we simplify
them a bunch and bring them under the GithubService fold. Methods that
acted like associations (like `issue.list_comments`) are removed,
because you should use the service to get that - but things that are
directly related to the response data or are custom actions (like
wipifying the title of an issue) are left alone.

We also remove a ton of Github related stuff in the Repo model and more
appropriately put it within the GithubService.

Lastly, we reimplement GithubNotificationMonitor, the only thing that
was using the OctokitWrappers, with the new interface.