Extract command logic from GHNM
This pulls all the private API in GHNM that processes the text of a
comment or issue and calls the method that does all the things for the
command, putting it in a new CommandDispatcher class.

This dispatcher will take in an issue to take action on, and the public
API #dispatch! will take an author and text from that author to take
action on. For now, it's just a cut and paste of the original GHNM
logic, but now we can abstract commands away to their own classes.

The whole point of this is to remove and clean up the logic around what
a command is and does, and separate it from Notification processing.
Then plugging in web hooks will be easy!