Pronto integration
Integration of pronto required to add few new gems and to update some of
the gems. The old method which provided the linter launching was
reworked to launch pronto which provide the linter launching. The output
of pronto is the result of all linters. The pronto result is transformed
into a structure which match the original one. This integration required
to change few tests and add new ones.

Closes #192