Branch status notification via Gitter
Added new worker which checks the latest two builds on a specified branch in a specified repository and send a notifying message to the specified gitter room. This message inform the room members about the status of the branch - branch is broken or fixed. The message depends on the latest two builds from which is the report created. Required specifications must be included in the file containing the settings.