Automated request of codeowner for PR review
This was necessary due to the fact that there is not possible to choose number zero for required approving reviews in the settings of branch protection. Codeowners are requested for a pull request review depending on specifications defined in CODEOWNERS file. This file includes files specifications and authors related to them. Files from the PR are compared if they match any of the pattern defined in the codeowners file. If there is match then a corresponding user is requested for a review.