Exception handling for unimplemented method in remove_reviewer
This is a temporary solution until merge of https://github.com/octokit/octokit.rb/pull/990.

Fix of undefined method `delete_pull_request_review_request' because of dependent pull request was not merged yet. It was solved via catching the NoMethodError exception. The call of this method will inform the requester of remove_reviewer command that this command is currently not working and it will perform no action (only posting a github comment with this informative text).