Add better regexp for puts
This is less of a "new feature" and more of a way we should consider
handling checking for `puts`.

In the current config by just looking for `puts` as a "word", we will
end up flagging it when it is in comments or just in an everyday
sentence (you know... since it is a 'word' in the English dictionary)

* * *

The alternative to this approach would be to add the following regexp
matcher preamble:

   /^([^#]+|)\b...\b/

To the "word" for of the content checker.  I figured this was more
invasive, but might be more flexible overall.