Remove active_bugzilla
Replaced with some inline http calls via Faraday.  Uses the REST API,
since the XMLRPC API has been deprecated and will be removed in future
BZ releases, and the bits needed for this app are small enough to not
warrant trying to get all of `active_bugzilla` up to to code with the
new API format.