[Gitter::Bot] Properly handle websocket auth
Gitter seems to have a custom auth "extension" with it's Bayeux
implementation.  References to that can be found here:

- https://github.com/gitterHQ/gitter-faye-client/blob/391345d6/gitter-faye.js#L10-L17
- https://gitlab.com/gitlab-org/gitter/webapp/blob/3cfaa1fb/server/web/bayeux/authenticator.js#L53-60

But basically we need to make sure when doing a handshake, we add a
`"ext" => {"token" => "MY_TOKEN"}` to the websocket message.  Once that
was in place, working with the faye endpoints worked as expected.