Add cross-repo-test command framework
Adds a command for running cross repo tests from a pull-request using
the bot.

In short this command, when executed, will create a PR to the configured
test_repo that will trigger a build on CI to run the specs for that
given change-set.

This commit only implements the command parsing logic, with the git
branch/commit/pull_request logic coming in a follow up commit.