[Linter::Base] Fix rubucop file permission check
When checking out the contents of the file to run linters against use
Rugged, the content was just copied to a new file without taking into
account the permissions that are set in the repository object.

This fix now performs a check and creates a file matching the
permissions found in the repo being linted.