[Branch] Add build failure concept
Use the Branch record to keep track of build failures that we find when
polling the Travis API.  The main purpose is to avoid spamming gitter
with extra notifications, but also can be used to identify that a branch
has previously been broken and now has been fixed.

Makes sense to also notify on the passing cases so others aren't taking
the time to investigate a failure when this has already been
investigated/fixed by another.