[StaleIssueMarker] Use GithubService.search_issues
This is a pretty large refactoring of the StaleIssueMarker, but makes
use of the `search_issues` api to reduce the number of API calls
required to run this worker.

As a side effect, it also simplifies each of the methods in the class so
they are a bit easier to digest.