[Command::AddLabel] Handle unassignable labels
Starting with the Jansa release, we are changing the backport process to
require code changes that are requesting a backport submit the
`jansa/yes?` first to get approval from the triage team.  This means
that setting the `jansa/yes` label (and similar labels from this release
forward) will not be allowed.

For the purposes of the bot, this means that any labels that have been
assigned a `*/yes` label (only `jansa/yes` for now), will be
automatically converted to `jansa/yes?`.