Add IsTeamMember concern
A helper module for checking if a user is a member of a team.

Currently only implements a check for a triage_team, but left the module
name vauge so it could be used for other team filtering as well.

The name of the triage team is currently configured in the settings as a
top level key `:triage_team_name`.  Most likely making the `org` be
configurable as well might make sense for developers working on
`miq_bot` in their own orgs.