[StaleIssueMarker] Bump stale_date to 3.months.ago
Note:  This is kind of a wash to test, as you probably can tell by the
specs (they didn't fail prior to the changes made).  This is because we
assume that we are getting anything prior to the `stale` date from the
query, so the test assumes the API query is working as expected.