[StaleIssueMarker] Refactor label checking / specs
Moved some code around so that `GithubService::Issue` was in charge of
checking if an issue was stale.

This caused some issues in the specs, since `double` wasn't now looking
for `labels`, but instead the newly created `stale?` method.

As a result, the specs no longer use double objects, but stub
`Sawyer::Resource` objects, which probably allows more of our app code
to be exercised.