[StaleIssueMarker] Handle recently stale + unmergeable
This update handles recently unmergeable + stale PRs in the following
fashion:

- Makes sure to add the `stale` label
- Closes immediately
- Doesn't do this in two steps (mark stale, then close)

The last step was a big motivator for the change here.

There is already a check to see if there are any PRs that exist that are
`unmergeable` + `stale` in existence, which can still happen with this
change in place (an already stale PR can become `unmergeable` during the
course of the 3 months of the stale label being applied, in which case
the bot should close it prior to the 3 month timeout in the second
search).

This change makes it so that if an already unmergeable PR exists that is
becoming stale, it will close it right away, and only comment once that
the PR is being closed (not twice, once for becoming stale, and again
once it is closed in the second loop).

* * *

As a result, a bit of refactoring was required, which just ensures we
are always attempting to add the stale label to PRs we are closing or
commenting on.  As a result, `mark_as_stale` now only does that
operation (ejects early if the label already exists), and
`comment_and_close` and `comment_as_stale` will call out to that prior
to their other operations.