[Vagrantfile] Use DNF ruby, not ruby-install
ruby-install was originally picked for the Vagrantfile to match what was
in the .travis.yml, and `dnf` only had 2.5.7 available. But based on
output from the bot in production:

> Checked commit NickLaMuro@b68ddde with ruby 2.5.7 ...
>
> Ref: https://github.com/ManageIQ/miq_bot/pull/497#issuecomment-624154539

We are using 2.5.7 there as well, so switching to use a `dnf` install to
speed up the the over all build time of the box.