[WIP] Use up to date linter configs
Right now, what ever is on the "fetched master" for the linter configs
will be what is used by Pronto, and not what is up to date with the
branch being checked against.

This "attempts" to generate a config based on the git blob for the
branch we are checking, but it is still not quite working.