Handle duplicate repo name where one is a PR and the other is bare
This handles the case where someone asks for the same repo multiple times,
but where one is a PR and the other is the bare repo.  This can even happen
accidentally, such as in the following case:

In ManageIQ/manageiq PR 1234, someone asks the bot to process /all, but
/all includes ManageIQ/manageiq.  The bot adds the PR to the test repo
list and now you have both the PR and the bare repo.  Additionally, it
will add the PR to the included list. manageiq-cross-repo will then see
that as an invalid combination because you can't test a bare repo while
also saying that it should include a PR version of that same repo.

This commit will prioritize PRs over bare repos when duplicates are
detected.